Show actual image instead of thumbnail in rss feed
If the media file is an image, showing just the thumbnail in the rss
feed means that one has to open a link to see the actual image, which
can be inconvenient. Ideally, one should be able to able to see the
image from the comfort of their rss reader.